requirements.txt
requests
jinja2
pillow